% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_matching.R
\name{ATC}
\alias{ATC}
\title{ATC of a matched dataset}
\usage{
ATC(AME_out)
}
\arguments{
\item{AME_out}{An object of class \code{AME}.}
}
\description{
\code{ATC} estimates the average treatment effect on the controls (ATC) of a
matched dataset.
}
\details{
The ATC is estimated as the CATE estimate across all control units in the
data.
}
\seealso{
\code{\link{CATE}}, \code{\link{ATE}}, \code{\link{ATT}}
}
