% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_matching.R
\name{ATE}
\alias{ATE}
\title{ATE of a matched dataset}
\usage{
ATE(AME_out)
}
\arguments{
\item{AME_out}{An object of class \code{AME}.}
}
\description{
\code{ATE} estimates the average treatment effect (ATE) of a matched dataset.
}
\details{
The ATE is estimated as the difference between the weighted treated and the
weighted control outcomes in the dataset. A unit's weight is the reciprocal
number of times it was matched. Unmatched units are omitted from the
computation.
}
