% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLAME_bit.R
\name{FLAME}
\alias{FLAME}
\title{Bit Vectors Implementation}
\usage{
FLAME(
  data,
  holdout = 0.1,
  C = 0.1,
  treatment_column_name = "treated",
  outcome_column_name = "outcome",
  PE_method = "elasticnet",
  user_PE_fun = NULL,
  PE_fun_params = NULL,
  repeats = FALSE,
  verbose = 2,
  want_pe = TRUE,
  want_bf = FALSE,
  early_stop_iterations = Inf,
  epsilon = 0.05,
  early_stop_un_c_frac = 0,
  early_stop_un_t_frac = 0,
  early_stop_pe = Inf,
  early_stop_bf = 0,
  missing_data = 0,
  missing_holdout = 0,
  n_data_imputations = 10,
  n_holdout_imputations = 10,
  opt = 0
)
}
\arguments{
\item{data}{Data to be matched. Either a dataframe or a path to a .csv file
to be read into a dataframe.}

\item{holdout}{Data to be used to compute predictive error. If a numeric
scalar between 0 and 1 (default = 0.1), that proportion of \code{data} will
made into a holdout set to compute predictive error and only the remaining
proportion of \code{data} will be matched. Otherwise, a dataframe or a path
to a csv file.}

\item{C}{A positive scalar denoting the tradeoff between BF and PE. Higher C
will prioritize more matches and lower C will prioritize not dropping
important covariates.}

\item{treatment_column_name}{A character with the name of the treatment
column in \code{data}.}

\item{outcome_column_name}{A character with the name of the outcome column
in \code{data}.}

\item{PE_method}{One of "elasticnet", .... Denotes the method to be used for
computation of PE.}

\item{repeats}{A logical scalar. If true, allows the same unit to be matched
multiple times, on different numbers of covariates.}

\item{verbose}{Controls output while FLAME is running. If 0, no output. If 1,
outputs the iteration every iteration. If 2, outputs the iteration and
number of unmatched units every 5 iterations. If 3, outputs the iteration
and number of unmatched units every 5 iterations.}

\item{want_pe}{A logical scalar. If TRUE, the predictive error (PE) at each
iteration will be returned.}

\item{want_bf}{A logical scalar. If TRUE, the balancing factor (BF) at each
iteration will be returned.}

\item{early_stop_iterations}{A nonnegative integer, denoting the number of
iterations of FLAME to be performed. If 0, one round of exact matching is
performed before terminating.}

\item{early_stop_un_c_frac}{A numeric value between 0 and 1 (inclusive). If
the proportion of control units that are unmatched falls below this value,
FLAME terminates.}

\item{early_stop_un_t_frac}{A numeric value between 0 and 1 (inclusive). If
the proportion of treatment units that are unmatched falls below this
value, FLAME terminates.}

\item{early_stop_pe}{A numeric value between 0 and 1 (inclusive).}

\item{early_stop_bf}{A numeric value between 0 and 1 (inclusive).}

\item{missing_data}{If 0, assumes no missingness in \code{data}. If
1, eliminates units with missingness from \code{data}. If 2, will not match
a unit on a covariate that it is missing. If 3, performs
\code{n_data_imputations} of MICE to impute the missing data.}

\item{missing_holdout}{If 0, assumes no missing data in
\code{holdout}. If 1, eliminates units with missingness from
\code{holdout}. If 2,  performs \code{n_holdout_imputations} of MICE
to impute the missing data.}

\item{n_data_imputations}{If \code{missing_data} = 2, performs
this many imputations of the missing data in \code{data} using MICE.}

\item{n_holdout_imputations}{If \code{missing_holdout} = 2,
performs this many imputations of the missing data in \code{holdout} using
MICE.}

\item{stop_unmatched_c}{A logical scalar. If TRUE, FLAME will terminate when
there are no more control units to be matched.}

\item{stop_unmatched_t}{A logical scalar. If TRUE, FLAME will terminate when
there are no more treated units to be matched.}
}
\description{
\code{FLAME_bit} runs the FLAME matching algorithm implemented using bit
vectors. The user is required to pass in \code{data}.
This is a longer description of what FLAME does.
}
\section{Parameters for Early Stopping}{

}

\section{Parameters for Missing Data}{

}

\examples{
data <- gen_data()
holdout <- gen_data()
FLAME_out <- FLAME(data = data, holdout = holdout)
}
