% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLAME_bit.R
\name{FLAME_bit}
\alias{FLAME_bit}
\title{Bit Vectors Implementation}
\usage{
FLAME_bit(
  data,
  treatment_column_name = "treated",
  outcome_column_name = "outcome",
  PE_method = "elasticnet",
  alpha = 0.1,
  C = 0.1,
  holdout = FALSE,
  repeats = FALSE,
  verbose = 2,
  want_pe = TRUE,
  early_stop_iterations = Inf,
  stop_unmatched_c = FALSE,
  early_stop_un_c_frac = 0.1,
  stop_unmatched_t = FALSE,
  early_stop_un_t_frac = 0.1,
  early_stop_pe = FALSE,
  early_stop_pe_frac = 0.01,
  want_bf = FALSE,
  early_stop_bf = FALSE,
  early_stop_bf_frac = 0.01,
  missing_data_replace = 0,
  missing_holdout_replace = 0,
  missing_holdout_imputations = 10,
  missing_data_imputations = 0,
  opt = 0
)
}
\arguments{
\item{data}{Data to be matched. Either a dataframe or a path to a .csv file
to be read into a dataframe.}

\item{treatment_column_name}{A character with the name of the treatment
column in \code{data}.}

\item{outcome_column_name}{A character with the name of the outcomee column
in \code{data}.}

\item{PE_method}{One of "elasticnet", .... Denotes the method to be used for
computation of PE.}

\item{alpha}{A positive scalar to be passed to glmnet for computation of PE
via elastic nets.}

\item{C}{A positive scalar denoting the tradeoff between BF and PE. Higher C
will prioritize more matches and lower C will prioritize not dropping
important covariates.}

\item{holdout}{Data to be used to compute predictive error. If FALSE
(default), 10% of \code{data} will be used for this purpose and only the
remaining 90% of \code{data} will be matched. Otherwise, a dataframe or a
path to a csv file.}

\item{repeats}{A logical scalar. If true, allows the same unit to be matched
multiple times, on different numbers of covariates.}

\item{verbose}{Controls output while FLAME is running. If 0, no output. If 1,
outputs the iteration every iteration. If 2, outputs the iteration and
number of unmatched units every 5 iterations. If 3, outputs the iteration
and number of unmatched units every 5 iterations.}

\item{want_pe}{A logical scalar. If TRUE, the predictive error (PE) at each
iteration will be returned.}

\item{early_stop_iterations}{A nonnegative integer, denoting the number of
iterations of FLAME to be performed. If 0, one round of exact matching is
performed before terminating.}

\item{stop_unmatched_c}{A logical scalar. If TRUE, FLAME will terminate when
there are no more control units to be matched.}

\item{early_stop_un_c_frac}{A numeric value between 0 and 1 (inclusive). If
the proportion of control units that are unmatched falls below this value,
FLAME terminates.}

\item{stop_unmatched_t}{A logical scalar. If TRUE, FLAME will terminate when
there are no more treated units to be matched.}

\item{early_stop_un_t_frac}{A numeric value between 0 and 1 (inclusive). If
the proportion of treatment units that are unmatched falls below this
value, FLAME terminates.}

\item{early_stop_pe}{A logical scalar. If TRUE....}

\item{early_stop_pe_frac}{A numeric value between 0 and 1 (inclusive).}

\item{want_bf}{A logical scalar. If TRUE, the balancing factor (BF) at each
iteration will be returned.}

\item{early_stop_bf}{A logical scalar. If TRUE....}

\item{early_stop_bf_frac}{A numeric value between 0 and 1 (inclusive).}

\item{missing_data_replace}{If 0, assumes no missingness in \code{data}. If
1, eliminates units with missingness from \code{data}. If 2, will not match
a unit on a covariate that it is missing. If 3, performs
\code{missing_data_imputations} of MICE to impute the missing data.}

\item{missing_holdout_replace}{If 0, assumes no missing data in
\code{holdout}. If 1, eliminates units with missingness from
\code{holdout}. If 2,  performs \code{missing_holdout_imputations} of MICE
to impute the missing data.}

\item{missing_holdout_imputations}{If \code{missing_holdout_replace} = 2,
performs this many imputations of the missing data in \code{holdout} using
MICE.}

\item{missing_data_imputations}{If \code{missing_data_replace} = 2, performs
this many imputations of the missing data in \code{data} using MICE.}
}
\description{
\code{FLAME_bit} runs the FLAME matching algorithm implemented using bit
vectors. The user is required to pass in \code{data}.
This is a longer description of what FLAME does.
}
\section{Parameters for Early Stopping}{

}

\section{Parameters for Missing Data}{

}

\examples{
data <- gen_data(100, 5)
FLAME_bit(data = data,)
}
