% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_matching.R
\name{mmg_of_unit}
\alias{mmg_of_unit}
\title{Returns the matched group of a given unit.}
\usage{
mmg_of_unit(unit, flame_obj, output_style = 1)
}
\arguments{
\item{unit}{An integer between 1 and \code{nrow(flame_obj)} for whom the
treatment effect is desired.}

\item{flame_obj}{An object returned by running \code{FLAME_bit}.}

\item{output_style}{If 1, only returns the values of covariates these units
were matched on. If 0, returns all covariate values.}
}
\description{
\code{mmg_of_unit} Returns the matched group of unit \code{unit}, as well as
the covariates of the units in the matched group.
}
