% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_matching.R
\name{te_of_unit}
\alias{te_of_unit}
\title{Compute the treatment effect of a given unit.}
\usage{
te_of_unit(unit, flame_obj)
}
\arguments{
\item{unit}{An integer between 1 and \code{nrow(flame_obj)} for whom the
treatment effect is desired.}

\item{flame_obj}{An object returned by running \code{FLAME_bit}.}
}
\description{
\code{te_of_unit} Computes the treatment effect of unit \code{unit} as equal
to the CATE of its matched group.
}
