% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_matching.R
\name{ATE}
\alias{ATE}
\title{Compute the ATE of a matched dataset.}
\usage{
ATE(flame_obj)
}
\arguments{
\item{flame_obj}{An object returned by running \code{FLAME_bit}.}
}
\description{
\code{ATE} Computes the average treatment effect (ATE) of a matched dataset
via the difference of the weighted treated and the weighted control outcomes
A unit's weight is the number of times it was matched.
}
