% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_matching.R
\name{ATE}
\alias{ATE}
\title{Compute the ATE of a matched dataset.}
\usage{
ATE(flame_obj)
}
\arguments{
\item{flame_obj}{An object returned by running \code{FLAME_bit}.}
}
\description{
\code{ATE} Computes the ATE of a matched dataset via a weighted (by size)
average of the CATEs of each matched group in the matched dataset.
}
