% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_matching.R
\name{ATE}
\alias{ATE}
\alias{ATT}
\alias{ATC}
\title{Average Treatment Effect estimates}
\usage{
ATE(ame_out)

ATT(ame_out)

ATC(ame_out)
}
\arguments{
\item{ame_out}{An object of class \code{ame}.}
}
\description{
These functions are deprecated and will be made defunct at a later release.
See \code{summary.ame} for average treatment effects estimates and their
variance.
}
\details{
\code{ATE}, \code{ATT}, and \code{ATC} estimate the average treatment effect
(ATE), average treatment effect on the treated (ATT), and average treatment
effect on the controls (ATC), respectively, of a matched dataset.

The ATE is estimated as the average CATE estimate across all matched units in
the data, while the ATT and ATC average only across matched treated or
matched control units, respectively.
}
\seealso{
\code{\link{CATE}}
}
