% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_matching.R
\name{ATT}
\alias{ATT}
\title{Compute the ATT of a matched dataset.}
\usage{
ATT(FLAME_out)
}
\arguments{
\item{FLAME_out}{An object returned by running \code{FLAME_bit}.}
}
\description{
\code{ATT} Computes the average treatment effect on the treated (ATT) of a matched dataset.
}
\details{
For each treated unit, its counterfactual outcome is estimated via the mean
outcome of control units in its matched group. This value is then averaged
across all treated units.
}
