% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_matching.R
\name{CATE}
\alias{CATE}
\title{Conditional Average Treatment Effects}
\usage{
CATE(units, AME_out)
}
\arguments{
\item{units}{A vector of indices for the units whose CATEs
are desired.}

\item{AME_out}{An object of class \code{AME}.}
}
\value{
A scalar numeric with the CATE estimate for \code{units}.
}
\description{
\code{CATE} returns an estimate of the conditional average treatment effect
for the subgroup defined by \code{units}.
}
\details{
This function returns a single CATE estimate for a subgroup of interest, as
defined by the units in \code{units}. This is computed by averaging the CATE
estimates of all specified units, weighted by the weights of their main
matched groups. The weight of a main matched group is defined as the sum of
the inverse weights of the constituent units. Unmatched units are omitted
from the computation. To retrieve CATE estimates for individual units or
matched groups, use \code{AME_out$CATE}.
}
\examples{
data <- gen_data()
holdout <- gen_data()
FLAME_out <- FLAME(data = data, holdout = holdout)
CATE(which(data$X1 == 1), FLAME_out)
}
\seealso{
\code{\link{FLAME}}, \code{\link{DAME}}
}
