% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_matching.R
\name{CATE}
\alias{CATE}
\title{Compute the treatment effect of a given unit.}
\usage{
CATE(units, FLAME_out)
}
\arguments{
\item{units}{A vector of integers between 1 and \code{nrow(FLAME_out$data)}
whose matched groups are desired.}

\item{FLAME_out}{An object returned by running \code{FLAME_bit}.}
}
\description{
\code{CATE} Computes the treatment effect of unit \code{unit} as equal
to the CATE of its matched group.
}
