% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_matching.R
\name{MG}
\alias{MG}
\title{Matched Groups}
\usage{
MG(units, AME_out, multiple = FALSE, index_only = FALSE)
}
\arguments{
\item{units}{A vector of indices for the units whose matched groups
are desired.}

\item{AME_out}{An object of class \code{AME}.}

\item{multiple}{A logical scalar. If \code{FALSE} (default), then \code{MG}
will only return the main matched group for each unit. See below for
details. Should not be set to \code{TRUE} if \code{AME_out} was generated
without replacement.}

\item{index_only}{A logical scalar. If \code{TRUE} then only the indices of
units in each matched group are returned.}
}
\value{
A list of length \code{length(units)}. For matched units, each entry is a
  data frame (if \code{multiple = FALSE}) or a list of data frames (if
  \code{multiple = TRUE}). For a given entry, these data frames are subsets
  of \code{data} passed to \code{\link{FLAME}} to generate \code{FLAME_out},
  whose rows correspond to the units in the matched group(s) of that entry
  and whose columns are the respective values of treatment, outcome, and
  covariates matched on. For unmatched units, the relevant list entries are
  \code{NULL}.
}
\description{
\code{MG} returns the matched groups of the supplied units.
}
\details{
\code{MG} returns the treatment, outcome, and covariates matched on for all
units in the relevant matched groups. If only the indices of units in the
matched groups are desired, \code{index_only} can be set to \code{TRUE}.

The \code{multiple} argument toggles whether only a unit's main matched group
(MMG) or all matched groups a unit is part of should be returned. A unit's
MMG contains its highest quality matches. Note that if the original call that
generated \code{AME_out} specified \code{replace = FALSE} then units only are
part of one matched group (their MMG) and this argument has no effect.
Otherwise, \code{multiple = TRUE} requests that \code{MG} returns all matched
groups and \code{multiple = FALSE} that only MMGs are returned.
}
\seealso{
\code{\link{FLAME}}, \code{\link{DAME}}
}
