% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_matching.R
\name{MG}
\alias{MG}
\title{Matched groups of units.}
\usage{
MG(units, flame_obj, multiple = FALSE)
}
\arguments{
\item{units}{A vector of integers between 1 and \code{nrow(flame_obj$data)}
whose matched groups are desired.}

\item{flame_obj}{An object returned by running \code{\link{FLAME}}.}

\item{multiple}{A logical scalar. If \code{FALSE} (default), then will only
return each unit's first matched group; else, will return all of them. See
below for details.}
}
\value{
A list of length \code{length(units)}. Each entry is a data frame (if
  \code{multiple = FALSE}) or a list of data frames (if \code{multiple =
  TRUE}). For a given entry, these data frames are subsets of \code{data}
  passed to \code{\link{FLAME}} to generate \code{flame_obj}, whose rows
  correspond to the units in the matched group(s) of that entry.

  Each entry thus always corresponds to the matched group(s) of the
  corresponding unit in \code{units}, but the structure of the entries
  depends on the value of \code{multiple} and on whether or not the call to
  \code{\link{FLAME}} that generated \code{flame_obj} specified \code{repeats =
  TRUE} or not. If it did, a unit may have several matched
  groups, corresponding to several sets of covariates it was matched on. In
  this case, \code{multiple = FALSE} requests that only the matched group
  with matches on the greatest number of covariates be returned and
  \code{multiple = TRUE} requests that all matched groups involving that unit
  be returned. If it did not, a unit only has one matched group
  that will be returned.

  The starred entries (*) in the returned data frames have the same meaning
  as in \code{flame_obj$data}, except for if both \code{multiple = TRUE} and
  \code{repeats = TRUE}. In this case, if \emph{all} units do not match on a given
  covariate, all entries of that covariate will be starred, even though a
  subset of the units may have matched on them.
}
\description{
Returns the matched groups of units \code{units}, as well as the covariates,
treatment, and outcome of the units in each matched group.
}
\details{
\code{MG} returns a list, each entry of which contains information about the
matched group(s) of the corresponding unit. If \code{\link{FLAME}} was run with
argument \code{repeats = FALSE} to generate \code{flame_obj}, \code{MG}
should be run with the default \code{multiple = FALSE} and will return
information about the matched group of each unit. If \code{link{FLAME}} was run
with argument \code{repeats = TRUE}, then \code{MG} may be run with either
\code{multiple = FALSE} or \code{multiple = TRUE}. In the first case, only
one matched group, that with matches on the greatest number of covariates,
will be returned. In the second case, all matched groups involving the
relevant unit will be returned.
}
\seealso{
\code{\link{FLAME}}
}
